MA-7: fixed Eugeniy's remarks on review
