+++ app/src/main/java/com/example/messengerAndroid/ui/myContacts/contactsViewModel/contract/UsersDataSelector.kt
@@ -3,5 +3,5 @@ package com.example.messengerAndroid.ui.myContacts.contactsViewModel.contract
 import com.example.messengerAndroid.data.contactsRepository.contactModel.User
 
 interface UsersDataSelector {
-    fun getUsers() : MutableList<User>
+    fun getUsers() : List<User>
 }
