+++ app/src/main/java/com/example/messengerAndroid/ui/myProfile/MyProfileActivity.kt
@@ -4,6 +4,7 @@ import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import androidx.activity.OnBackPressedCallback
+import com.example.messengerAndroid.R
 import com.example.messengerAndroid.data.preferences.SharedPreferencesHelper
 import com.example.messengerAndroid.databinding.ActivityMyProfileBinding
 import com.example.messengerAndroid.extensions.cropPhoto
@@ -18,7 +19,7 @@ class MyProfileActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         binding = ActivityMyProfileBinding.inflate(layoutInflater)
-        binding.imageViewPicture.cropPhoto()
+        binding.imageViewPicture.cropPhoto(R.drawable.babyyoda)
         binding.textViewName.text = intent.getStringExtra(USER_NAME_KEY)
         setContentView(binding.root)
